Moved image fetching to custom games
