+++ app/src/main/java/app/gamenative/ui/screen/library/appscreen/GOGAppScreen.kt
@@ -436,9 +436,6 @@ class GOGAppScreen : BaseAppScreen() {
             },
         )
     }
-    override fun getGameFolderPathForImageFetch(context: Context, libraryItem: LibraryItem): String? {
-        return null // GOG Stores full URLs in their database entry.
-    }
 
     override fun observeGameState(
         context: Context,
